Use double instead of float, unless explicitely necessary

+++ include/BeadingStrategy/LimitedBeadingStrategy.h
@@ -39,7 +39,7 @@ public:
 
     coord_t getTransitioningLength(coord_t lower_bead_count) const override;
 
-    float getTransitionAnchorPos(coord_t lower_bead_count) const override;
+    double getTransitionAnchorPos(coord_t lower_bead_count) const override;
 
 protected:
     const coord_t max_bead_count;
