+++ src/BeadingStrategy/DistributedBeadingStrategy.cpp
@@ -1,23 +1,22 @@
 // Copyright (c) 2022 Ultimaker B.V.
 // CuraEngine is released under the terms of the AGPLv3 or higher.
-#include <numeric>
 #include "BeadingStrategy/DistributedBeadingStrategy.h"
+
+#include <numeric>
 
 namespace cura
 {
 
-DistributedBeadingStrategy::DistributedBeadingStrategy
-(
+DistributedBeadingStrategy::DistributedBeadingStrategy(
     const coord_t optimal_width,
     const coord_t default_transition_length,
     const AngleRadians transitioning_angle,
     const Ratio wall_split_middle_threshold,
     const Ratio wall_add_middle_threshold,
-    const int distribution_radius
-) :
-    BeadingStrategy(optimal_width, wall_split_middle_threshold, wall_add_middle_threshold, default_transition_length, transitioning_angle)
+    const int distribution_radius)
+    : BeadingStrategy(optimal_width, wall_split_middle_threshold, wall_add_middle_threshold, default_transition_length, transitioning_angle)
 {
-    if(distribution_radius >= 2)
+    if (distribution_radius >= 2)
     {
         one_over_distribution_radius_squared = 1.0f / (distribution_radius - 1) * 1.0f / (distribution_radius - 1);
     }
@@ -36,25 +35,25 @@ DistributedBeadingStrategy::Beading DistributedBeadingStrategy::compute(coord_t 
     if (bead_count > 2)
     {
         const coord_t to_be_divided = thickness - bead_count * optimal_width;
-        const float middle = static_cast<float>(bead_count - 1) / 2;
+        const double middle = static_cast<double>(bead_count - 1) / 2;
 
         const auto getWeight = [middle, this](coord_t bead_idx)
         {
-            const float dev_from_middle = bead_idx - middle;
-            return std::max(0.0f, 1.0f - one_over_distribution_radius_squared * dev_from_middle * dev_from_middle);
+            const double dev_from_middle = bead_idx - middle;
+            return std::max(0.0, 1.0 - one_over_distribution_radius_squared * dev_from_middle * dev_from_middle);
         };
 
-        std::vector<float> weights;
+        std::vector<double> weights;
         weights.resize(bead_count);
         for (coord_t bead_idx = 0; bead_idx < bead_count; bead_idx++)
         {
             weights[bead_idx] = getWeight(bead_idx);
         }
 
-        const float total_weight = std::accumulate(weights.cbegin(), weights.cend(), 0.f);
+        const double total_weight = std::accumulate(weights.cbegin(), weights.cend(), 0.0);
         for (coord_t bead_idx = 0; bead_idx < bead_count; bead_idx++)
         {
-            const float weight_fraction = weights[bead_idx] / total_weight;
+            const double weight_fraction = weights[bead_idx] / total_weight;
             const coord_t splitup_left_over_weight = to_be_divided * weight_fraction;
             const coord_t width = optimal_width + splitup_left_over_weight;
             if (bead_idx == 0)
